requirements.txt
Flask
Flask-Session
psycopg2
SQLAlchemy
Flask-Login
sqlalchemy.orm
wtforms
passlib.hash
functools
simplejson